refactor(agent): group the missing-reasoning incident into one watch
Four fields on Agent tracked one provider incident: whether it is active,
whether the persisted state records it, the healthy-round streak, and a resolve
timestamp whose write failed. Only observeMissingToolCallReasoning ever moved
them, and it moved them together in twelve places across eighty lines — an
interleaving that reads as four independent flags while only a few of their
combinations are reachable.

missingReasoningWatch names the four with their roles rather than their history:
warnedMissingToolCallReasoning becomes active, missingReasoningWarnStateChecked
becomes stateRecorded. The legacy names were kept for the on-disk contract,
which the persisted state (missingReasoningWarnState, untouched here) still
owns; the in-memory view has no such obligation.

The type and its only user move to their own file, which also takes 87 lines out
of agent.go's 3194.

One behaviour is deliberately preserved rather than tidied: SetSession clears
three of the four and leaves pendingResolveAt alone. Zeroing the struct would
have been the obvious write and would have silently dropped an unwritten resolve
watermark on every session switch — the watermark belongs to a provider
configuration, not to the conversation being replaced. The reset is explicit and
says so.

Second bucket under the struct-state ratchet: Agent 56 -> 52 scalars, baseline
tightened to match (agent.go 44 -> 40, total 156 -> 152). The file-size budget
also drops 2404 -> 2317 for the 87 lines moved out; the 10-unit drift already
recorded there is left as-is, since correcting it belongs to whatever PR takes
on the baseline drift as a whole.
